fix(test fetch_data) split in 2 files
